gio/glib: manually implement DBusProxy::g-signal
As the signal details were exposed only in 2.72
Also add missing v2_72 guards where needed
