+++ gio/Gir.toml
@@ -272,7 +272,7 @@ status = "generate"
         string_type = "os_string"
     [[object.signal]]
     name = "launch-started"
-    version = "2.70" # has the annotation upstream no idea why it doesn't appear in the gir file
+    version = "2.72" # has the annotation upstream no idea why it doesn't appear in the gir file
 
 [[object]]
 name = "Gio.Application"
@@ -528,6 +528,7 @@ status = "generate"
 name = "Gio.DBusProxy"
 status = "generate"
 concurrency = "send+sync"
+manual_traits = ["DBusProxyExtManual"]
     [[object.signal]]
     name = "g-properties-changed"
     # libc::c_char vs str
