+++ gio/sys/Cargo.toml
@@ -32,6 +32,7 @@ dox = []
 v2_66 = ["v2_64"]
 v2_68 = ["v2_66"]
 v2_70 = ["v2_68"]
+v2_72 = ["v2_70"]
 
 [lib]
 name = "gio_sys"
