+++ glib/Cargo.toml
@@ -51,6 +51,7 @@ v2_64 = ["v2_62", "ffi/v2_64"]
 v2_66 = ["v2_64", "ffi/v2_66", "gobject_ffi/v2_66"]
 v2_68 = ["v2_66", "ffi/v2_68", "gobject_ffi/v2_68"]
 v2_70 = ["v2_68", "ffi/v2_70", "gobject_ffi/v2_70"]
+v2_72 = ["v2_70", "ffi/v2_72"]
 log = ["rs-log"]
 log_macros = ["log"]
 dox = ["ffi/dox", "gobject_ffi/dox", "log_macros"]
