+++ glib/Gir.toml
@@ -510,7 +510,7 @@ status = "generate"
     ignore = true
     [[object.function]]
     name = "get_user_state_dir"
-    version = "2.70"
+    version = "2.72"
 
 [[object]]
 name = "GLib.Checksum"
