+++ glib/src/auto/functions.rs
@@ -423,8 +423,8 @@ pub fn user_special_dir(directory: UserDirectory) -> Option<std::path::PathBuf> 
     unsafe { from_glib_none(ffi::g_get_user_special_dir(directory.into_glib())) }
 }
 
-#[cfg(any(feature = "v2_70", feature = "dox"))]
-#[cfg_attr(feature = "dox", doc(cfg(feature = "v2_70")))]
+#[cfg(any(feature = "v2_72", feature = "dox"))]
+#[cfg_attr(feature = "dox", doc(cfg(feature = "v2_72")))]
 #[doc(alias = "g_get_user_state_dir")]
 #[doc(alias = "get_user_state_dir")]
 pub fn user_state_dir() -> crate::GString {
