+++ glib/sys/Gir.toml
@@ -118,6 +118,9 @@ status = "generate"
     name = "win32_get_package_installation_directory_of_module"
     manual = true
     cfg_condition = "windows"
+    [[object.function]]
+    name = "get_user_state_dir"
+    version = "2.72"
 
 [[object]]
 name = "GLib.Dir"
